LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/animations/font-variation-settings-interpolation-expected.txt


PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (-0.5) should be ['test' 5]
PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (0) should be ['test' 10]
PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (0.3) should be ['test' 13]
PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (0.7) should be ['test' 17]
PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (1) should be ['test' 20]
PASS CSS Transitions: property <font-variation-settings> from neutral to ["test" 20] at (1.5) should be ['test' 25]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (-0.5) should be ['test' 5]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (0) should be ['test' 10]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (0.3) should be ['test' 13]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (0.7) should be ['test' 17]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (1) should be ['test' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from neutral to ["test" 20] at (1.5) should be ['test' 25]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (-0.5) should be ['test' 5]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (0) should be ['test' 10]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (0.3) should be ['test' 13]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (0.7) should be ['test' 17]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (1) should be ['test' 20]
PASS CSS Animations: property <font-variation-settings> from neutral to ["test" 20] at (1.5) should be ['test' 25]
PASS Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (-0.5) should be ['test' 5]
FAIL Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (0) should be ['test' 10] assert_array_equals: expected property 0 to be "\"test\" 10" but got "\"test\" 5" (expected array ["\"test\" 10"] got ["\"test\" 5"])
FAIL Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (0.3) should be ['test' 13] assert_array_equals: expected property 0 to be "\"test\" 13" but got "\"test\" 9.5" (expected array ["\"test\" 13"] got ["\"test\" 9.5"])
FAIL Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (0.7) should be ['test' 17] assert_array_equals: expected property 0 to be "\"test\" 17" but got "\"test\" 16.85" (expected array ["\"test\" 17"] got ["\"test\" 16.85"])
PASS Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (1) should be ['test' 20]
FAIL Web Animations: property <font-variation-settings> from neutral to ["test" 20] at (1.5) should be ['test' 25] assert_array_equals: expected property 0 to be "\"test\" 25" but got "\"test\" 20" (expected array ["\"test\" 25"] got ["\"test\" 20"])
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (-0.3) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 24 "
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (0) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 30 "
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (0.3) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 36 "
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (0.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 40 "
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (0.6) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 42 "
PASS CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (1) should be ['test' 50]
FAIL CSS Transitions: property <font-variation-settings> from [initial] to ['test' 50] at (1.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 60 "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (-0.3) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 24 "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (0) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 30 "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (0.3) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 36 "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (0.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 40 "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (0.6) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 42 "
PASS CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (1) should be ['test' 50]
FAIL CSS Transitions with transition: all: property <font-variation-settings> from [initial] to ['test' 50] at (1.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 60 "
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (-0.3) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" - 2 "
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" 10 "
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.3) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" 22 "
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 30 "
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.6) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 34 "
PASS CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (1) should be ['test' 50]
FAIL CSS Animations: property <font-variation-settings> from [initial] to ['test' 50] at (1.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 70 "
FAIL Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (-0.3) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" - 2 "
FAIL Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" - 2 "
FAIL Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.3) should be [initial] assert_equals: expected "\" test \" 30 " but got "\" test \" 13.6 "
FAIL Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.5) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 31.8 "
FAIL Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (0.6) should be ['test' 50] assert_equals: expected "\" test \" 50 " but got "\" test \" 42.72 "
PASS Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (1) should be ['test' 50]
PASS Web Animations: property <font-variation-settings> from [initial] to ['test' 50] at (1.5) should be ['test' 50]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (-0.5) should be ['test' 35]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (0) should be ['test' 30]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (0.3) should be ['test' 27]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (0.7) should be ['test' 23]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (1) should be ['test' 20]
PASS CSS Transitions: property <font-variation-settings> from [inherit] to ['test' 20] at (1.5) should be ['test' 15]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (-0.5) should be ['test' 35]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (0) should be ['test' 30]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (0.3) should be ['test' 27]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (0.7) should be ['test' 23]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (1) should be ['test' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [inherit] to ['test' 20] at (1.5) should be ['test' 15]
FAIL CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (-0.5) should be ['test' 35] assert_array_equals: expected property 0 to be "\"test\" 35" but got "\"test\" 5" (expected array ["\"test\" 35"] got ["\"test\" 5"])
FAIL CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0) should be ['test' 30] assert_array_equals: expected property 0 to be "\"test\" 30" but got "\"test\" 10" (expected array ["\"test\" 30"] got ["\"test\" 10"])
FAIL CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0.3) should be ['test' 27] assert_array_equals: expected property 0 to be "\"test\" 27" but got "\"test\" 13" (expected array ["\"test\" 27"] got ["\"test\" 13"])
FAIL CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0.7) should be ['test' 23] assert_array_equals: expected property 0 to be "\"test\" 23" but got "\"test\" 17" (expected array ["\"test\" 23"] got ["\"test\" 17"])
PASS CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (1) should be ['test' 20]
FAIL CSS Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (1.5) should be ['test' 15] assert_array_equals: expected property 0 to be "\"test\" 15" but got "\"test\" 25" (expected array ["\"test\" 15"] got ["\"test\" 25"])
FAIL Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (-0.5) should be ['test' 35] assert_array_equals: expected property 0 to be "\"test\" 35" but got "\"test\" 5" (expected array ["\"test\" 35"] got ["\"test\" 5"])
FAIL Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0) should be ['test' 30] assert_array_equals: expected property 0 to be "\"test\" 30" but got "\"test\" 5" (expected array ["\"test\" 30"] got ["\"test\" 5"])
FAIL Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0.3) should be ['test' 27] assert_array_equals: expected property 0 to be "\"test\" 27" but got "\"test\" 9.5" (expected array ["\"test\" 27"] got ["\"test\" 9.5"])
FAIL Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (0.7) should be ['test' 23] assert_array_equals: expected property 0 to be "\"test\" 23" but got "\"test\" 16.85" (expected array ["\"test\" 23"] got ["\"test\" 16.85"])
PASS Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (1) should be ['test' 20]
FAIL Web Animations: property <font-variation-settings> from [inherit] to ['test' 20] at (1.5) should be ['test' 15] assert_array_equals: expected property 0 to be "\"test\" 15" but got "\"test\" 20" (expected array ["\"test\" 15"] got ["\"test\" 20"])
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (-0.5) should be ['test' 35]
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (0) should be ['test' 30]
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (0.3) should be ['test' 27]
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (0.7) should be ['test' 23]
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (1) should be ['test' 20]
PASS CSS Transitions: property <font-variation-settings> from [unset] to ['test' 20] at (1.5) should be ['test' 15]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (-0.5) should be ['test' 35]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (0) should be ['test' 30]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (0.3) should be ['test' 27]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (0.7) should be ['test' 23]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (1) should be ['test' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from [unset] to ['test' 20] at (1.5) should be ['test' 15]
FAIL CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (-0.5) should be ['test' 35] assert_array_equals: expected property 0 to be "\"test\" 35" but got "\"test\" 5" (expected array ["\"test\" 35"] got ["\"test\" 5"])
FAIL CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0) should be ['test' 30] assert_array_equals: expected property 0 to be "\"test\" 30" but got "\"test\" 10" (expected array ["\"test\" 30"] got ["\"test\" 10"])
FAIL CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0.3) should be ['test' 27] assert_array_equals: expected property 0 to be "\"test\" 27" but got "\"test\" 13" (expected array ["\"test\" 27"] got ["\"test\" 13"])
FAIL CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0.7) should be ['test' 23] assert_array_equals: expected property 0 to be "\"test\" 23" but got "\"test\" 17" (expected array ["\"test\" 23"] got ["\"test\" 17"])
PASS CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (1) should be ['test' 20]
FAIL CSS Animations: property <font-variation-settings> from [unset] to ['test' 20] at (1.5) should be ['test' 15] assert_array_equals: expected property 0 to be "\"test\" 15" but got "\"test\" 25" (expected array ["\"test\" 15"] got ["\"test\" 25"])
FAIL Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (-0.5) should be ['test' 35] assert_array_equals: expected property 0 to be "\"test\" 35" but got "\"test\" 5" (expected array ["\"test\" 35"] got ["\"test\" 5"])
FAIL Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0) should be ['test' 30] assert_array_equals: expected property 0 to be "\"test\" 30" but got "\"test\" 5" (expected array ["\"test\" 30"] got ["\"test\" 5"])
FAIL Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0.3) should be ['test' 27] assert_array_equals: expected property 0 to be "\"test\" 27" but got "\"test\" 9.5" (expected array ["\"test\" 27"] got ["\"test\" 9.5"])
FAIL Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (0.7) should be ['test' 23] assert_array_equals: expected property 0 to be "\"test\" 23" but got "\"test\" 16.85" (expected array ["\"test\" 23"] got ["\"test\" 16.85"])
PASS Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (1) should be ['test' 20]
FAIL Web Animations: property <font-variation-settings> from [unset] to ['test' 20] at (1.5) should be ['test' 15] assert_array_equals: expected property 0 to be "\"test\" 15" but got "\"test\" 20" (expected array ["\"test\" 15"] got ["\"test\" 20"])
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (-0.3) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (0) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (0.3) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (0.5) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (0.6) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (1) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to [normal] at (1.5) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (-0.3) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (0) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (0.3) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (0.5) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (0.6) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (1) should be [normal]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to [normal] at (1.5) should be [normal]
FAIL CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (-0.3) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.3) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.5) should be [normal]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.6) should be [normal]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (1) should be [normal]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to [normal] at (1.5) should be [normal]
FAIL Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (-0.3) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.3) should be ['test' 20] assert_equals: expected "\" test \" 20 " but got "normal "
PASS Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.5) should be [normal]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (0.6) should be [normal]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (1) should be [normal]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to [normal] at (1.5) should be [normal]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (-0.5) should be ['test' 15]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0) should be ['test' 20]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.3) should be ['test' 23]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.7) should be ['test' 27]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1) should be ['test' 30]
PASS CSS Transitions: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1.5) should be ['test' 35]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (-0.5) should be ['test' 15]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0) should be ['test' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.3) should be ['test' 23]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.7) should be ['test' 27]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1) should be ['test' 30]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1.5) should be ['test' 35]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (-0.5) should be ['test' 15]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0) should be ['test' 20]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.3) should be ['test' 23]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.7) should be ['test' 27]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1) should be ['test' 30]
PASS CSS Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1.5) should be ['test' 35]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (-0.5) should be ['test' 15]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0) should be ['test' 20]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.3) should be ['test' 23]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (0.7) should be ['test' 27]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1) should be ['test' 30]
PASS Web Animations: property <font-variation-settings> from ['test' 20] to ['test' 30] at (1.5) should be ['test' 35]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.5) should be ['aaaa' -5, 'bbbb' 5, 'cccc' 15]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 3, 'bbbb' 13, 'cccc' 23]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.7) should be ['aaaa' 7, 'bbbb' 17, 'cccc' 27]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 15, 'bbbb' 25, 'cccc' 35]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.5) should be ['aaaa' -5, 'bbbb' 5, 'cccc' 15]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 3, 'bbbb' 13, 'cccc' 23]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.7) should be ['aaaa' 7, 'bbbb' 17, 'cccc' 27]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 15, 'bbbb' 25, 'cccc' 35]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.5) should be ['aaaa' -5, 'bbbb' 5, 'cccc' 15]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 3, 'bbbb' 13, 'cccc' 23]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.7) should be ['aaaa' 7, 'bbbb' 17, 'cccc' 27]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 15, 'bbbb' 25, 'cccc' 35]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.5) should be ['aaaa' -5, 'bbbb' 5, 'cccc' 15]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 3, 'bbbb' 13, 'cccc' 23]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.7) should be ['aaaa' 7, 'bbbb' 17, 'cccc' 27]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 15, 'bbbb' 25, 'cccc' 35]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (-0.5) should be ['aaaa' -15, 'bbbb' 5, 'cccc' 25]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.3) should be ['aaaa' 9, 'bbbb' 13, 'cccc' 17]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.7) should be ['aaaa' 21, 'bbbb' 17, 'cccc' 13]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1) should be ['aaaa' 30, 'bbbb' 20, 'cccc' 10]
PASS CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1.5) should be ['aaaa' 45, 'bbbb' 25, 'cccc' 5]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (-0.5) should be ['aaaa' -15, 'bbbb' 5, 'cccc' 25]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.3) should be ['aaaa' 9, 'bbbb' 13, 'cccc' 17]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.7) should be ['aaaa' 21, 'bbbb' 17, 'cccc' 13]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1) should be ['aaaa' 30, 'bbbb' 20, 'cccc' 10]
PASS CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1.5) should be ['aaaa' 45, 'bbbb' 25, 'cccc' 5]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (-0.5) should be ['aaaa' -15, 'bbbb' 5, 'cccc' 25]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.3) should be ['aaaa' 9, 'bbbb' 13, 'cccc' 17]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.7) should be ['aaaa' 21, 'bbbb' 17, 'cccc' 13]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1) should be ['aaaa' 30, 'bbbb' 20, 'cccc' 10]
PASS CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1.5) should be ['aaaa' 45, 'bbbb' 25, 'cccc' 5]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (-0.5) should be ['aaaa' -15, 'bbbb' 5, 'cccc' 25]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.3) should be ['aaaa' 9, 'bbbb' 13, 'cccc' 17]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (0.7) should be ['aaaa' 21, 'bbbb' 17, 'cccc' 13]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1) should be ['aaaa' 30, 'bbbb' 20, 'cccc' 10]
PASS Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['cccc' 10, 'bbbb' 20, 'aaaa' 30] at (1.5) should be ['aaaa' 45, 'bbbb' 25, 'cccc' 5]
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.6) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.6) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.6) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (-0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (0.6) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10] to ['aaaa' 10, 'bbbb' 20, 'cccc' 30] at (1.5) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (-0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.6) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (-0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.3) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.6) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (-0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.6) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (-0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.3) should be ['aaaa' 10, 'bbbb' 20, 'cccc' 30] assert_equals: expected "\" cccc \" 30 , \" bbbb \" 20 , \" aaaa \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (0.6) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 10, 'bbbb' 20, 'cccc' 30] to ['aaaa' 0, 'bbbb' 10] at (1.5) should be ['aaaa' 0, 'bbbb' 10] assert_equals: expected "\" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (-0.3) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.3) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.6) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (-0.3) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.3) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.6) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (-0.3) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.3) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.6) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (-0.3) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.3) should be ['aaaa' 0, 'bbbb' 10, 'cccc' 20] assert_equals: expected "\" cccc \" 20 , \" bbbb \" 10 , \" aaaa \" 0 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (0.6) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 0, 'bbbb' 10, 'cccc' 20] to ['dddd' 10, 'eeee' 20, 'ffff' 30] at (1.5) should be ['dddd' 10, 'eeee' 20, 'ffff' 30] assert_equals: expected "\" ffff \" 30 , \" eeee \" 20 , \" dddd \" 10 " but got "normal "
FAIL CSS Transitions: property <font-variation-settings> from ['aaaa' 30, 'bbbb' 20] to ['aaaa' 20, 'bbbb' 30] at (3.40282e+38) should be ['aaaa' -3.40282e+38, 'bbbb' 3.40282e+38] assert_array_equals: expected property 0 to be "\"aaaa\" -3.40282e+38" but got "\"aaaa\" -Infinity" (expected array ["\"aaaa\" -3.40282e+38", "\"bbbb\" 3.40282e+38"] got ["\"aaaa\" -Infinity", "\"bbbb\" Infinity"])
FAIL CSS Transitions with transition: all: property <font-variation-settings> from ['aaaa' 30, 'bbbb' 20] to ['aaaa' 20, 'bbbb' 30] at (3.40282e+38) should be ['aaaa' -3.40282e+38, 'bbbb' 3.40282e+38] assert_array_equals: expected property 0 to be "\"aaaa\" -3.40282e+38" but got "\"aaaa\" -Infinity" (expected array ["\"aaaa\" -3.40282e+38", "\"bbbb\" 3.40282e+38"] got ["\"aaaa\" -Infinity", "\"bbbb\" Infinity"])
FAIL CSS Animations: property <font-variation-settings> from ['aaaa' 30, 'bbbb' 20] to ['aaaa' 20, 'bbbb' 30] at (3.40282e+38) should be ['aaaa' -3.40282e+38, 'bbbb' 3.40282e+38] assert_array_equals: expected property 0 to be "\"aaaa\" -3.40282e+38" but got "\"aaaa\" -Infinity" (expected array ["\"aaaa\" -3.40282e+38", "\"bbbb\" 3.40282e+38"] got ["\"aaaa\" -Infinity", "\"bbbb\" Infinity"])
FAIL Web Animations: property <font-variation-settings> from ['aaaa' 30, 'bbbb' 20] to ['aaaa' 20, 'bbbb' 30] at (3.40282e+38) should be ['aaaa' -3.40282e+38, 'bbbb' 3.40282e+38] assert_array_equals: expected property 0 to be "\"aaaa\" -3.40282e+38" but got "\"aaaa\" -Infinity" (expected array ["\"aaaa\" -3.40282e+38", "\"bbbb\" 3.40282e+38"] got ["\"aaaa\" -Infinity", "\"bbbb\" Infinity"])
